+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Unreleased
 Breaking API Changes
 ~~~~~~~~~~~~~~~~~~~~
 
+* finagle-base-http: Removed the methods `setStatusCode` and `getStatusCode` from
+  `c.t.f.http.Response` which have been deprecated since 2017. ``PHAB_ID=D326326``
+
 * finagle-core: A long deprecated `c.t.f.builder.ServerBuilder#build` method, which
   takes a `Service[Req, Rep]` and a not fully specified config as params, has been
   removed. ``PHAB_ID=D330348``